pyproject.toml
[tool.poetry]
name = "blog-contents"
version = "0.1.0"
description = ""
authors = ["ayu0616 <ayumuogawa0616@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.20240311"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
